+++ src/lib/Game.h
@@ -1,0 +1,15 @@
+#ifndef GAME_H
+#define GAME_H
+
+#include "Entity.h"
+
+
+class Game {
+    
+    public:
+        void initGame();
+        void restart();
+};
+
+
+#endif
